requirements-i.txt
accelerate==0.34.2
addict==2.4.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
anls==0.0.2
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.1.1
cryptography==43.0.1
cssselect2==0.7.0
cycler==0.12.1
datasets==3.0.1
defusedxml==0.7.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.16
einops==0.8.0
exceptiongroup==1.2.2
fastapi==0.112.4
filelock==3.16.1
fire==0.6.0
flash-attention==1.0.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gguf==0.10.0
google-api-core==2.20.0
google-api-python-client==2.148.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
h11==0.14.0
h5py==3.11.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.25.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
interegular==0.3.3
Jinja2==3.1.4
jiter==0.6.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
langdetect==1.0.9
lark==1.2.2
llama_stack==0.0.1a5
llvmlite==0.41.1
lm-format-enforcer==0.10.6
lmdeploy==0.5.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mistral_common==1.4.4
mistral_inference==0.0.0
mmengine-lite==0.10.4
mplcursors==0.5.3
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.18.6
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.1
ninja==1.11.1.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnx==1.16.2
openai==1.51.2
opencv-python==4.10.0.84
outlines==0.0.46
packaging==24.1
pandas==2.0.3
partial-json-parser==0.2.1.1.post4
pathlib==1.0.1
peft==0.11.1
pickle5==0.0.11
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.0
propcache==0.2.0
proto-plus==1.24.0
protobuf==3.20.3
psutil==6.0.0
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post5
pycountry==24.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydyf==0.11.0
Pygments==2.18.0
pynvml==11.5.3
pyparsing==3.1.4
pypdfium2==4.30.0
pyphen==0.16.0
python-dateutil==2.9.0.post0
python-doctr==0.8.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
rapidfuzz==3.9.7
ray==2.10.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
scipy==1.10.1
sentencepiece==0.2.0
shapely==2.0.6
shortuuid==1.0.13
simple-parsing==0.1.6
six==1.16.0
sniffio==1.3.1
starlette==0.38.5
sympy==1.13.2
termcolor==2.4.0
tf2onnx==1.16.1
tiktoken==0.7.0
timm==1.0.9
tinycss2==1.3.0
tokenizers==0.20.0
tomli==2.0.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.5
transformers==4.45.1
triton==2.3.1
typing_extensions==4.12.2
tzdata==2024.2
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.20.0
vllm==0.6.2
watchfiles==0.24.0
weasyprint==61.2
webencodings==0.5.1
websockets==13.1
xformers==0.0.27.post2
xxhash==3.5.0
yapf==0.40.2
yarl==1.14.0
ZDrive==2.1.5
zipp==3.20.2
zopfli==0.2.3
